Muevo el json de endpoints a un archivo separado
